cleanup stage 1, preparing to make gem
