+++ .gitignore
@@ -1,12 +1,13 @@
-# env.json
-
-node_modules
-puppeteer_cache
-dist
+# files generated by the script
 
 error.log
 market_cache
 bdo_codex_cache
+
+# for my notes and stuff
+npc_item_index.txt
+*.md
+!readme.md
 
 .idea
