--- env.mjs
@@ -1,69 +1,0 @@
-const COOKIE =
-  '__RequestVerificationToken=0q_pyZ8OMkALLxxv_pj6ty10hXKTDq3Sl_1NAbV5WYhUAbNvcurT6GwSUDmBRnaZVyEmd0lcTlR3I6X7_cXkAwH-nGQG_G2E6MQZPNcfNd41'
-
-export default {
-  RECIPE_FILE_NAME: 'recipeCache.json',
-  RVT: '__RequestVerificationToken=MGf54RSTVQkYrPo4Dvaf9IOuSilYwQIzWfdBI4PhFrV1l1o_e0BRmf78J6I7pVn_gcqj5DAv-G3MfFiWoUqFE3R62-Kxdcm4CgkRGd7oq7Y1',
-  COOKIE,
-  BATCH_SIZE: 0,
-  HIDE_UNPROFITABLE_RECIPES: true,
-  HIDE_OUT_OF_STOCK: true,
-  ROOT_URL: 'https://na-trade.naeu.playblackdesert.com/Home',
-  WORLD_MARKET_LIST: '/GetWorldMarketList',
-  MARKET_SUB_LIST: '/GetWorldMarketSubList',
-  MARKET_SEARCH_LIST: '/GetWorldMarketSearchList',
-  MARKET_SELL_BUY_INFO: '/GetItemSellBuyInfo',
-  REQUEST_OPTS: {
-    method: 'post',
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      'User-Agent':
-        'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36',
-      Cookie: COOKIE,
-    },
-  },
-  AXIOS_HEADERS: {
-    'user-agent':
-      'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36',
-    dnt: 1,
-    'x-requested-with': 'XMLHttpRequest',
-  },
-  PUPPETEER_ARGS: [
-    '--autoplay-policy=user-gesture-required',
-    '--disable-background-networking',
-    '--disable-background-timer-throttling',
-    '--disable-backgrounding-occluded-windows',
-    '--disable-breakpad',
-    '--disable-client-side-phishing-detection',
-    '--disable-component-update',
-    '--disable-default-apps',
-    '--disable-dev-shm-usage',
-    '--disable-domain-reliability',
-    '--disable-extensions',
-    '--disable-features=AudioServiceOutOfProcess',
-    '--disable-hang-monitor',
-    '--disable-ipc-flooding-protection',
-    '--disable-notifications',
-    '--disable-offer-store-unmasked-wallet-cards',
-    '--disable-popup-blocking',
-    '--disable-print-preview',
-    '--disable-prompt-on-repost',
-    '--disable-renderer-backgrounding',
-    '--disable-setuid-sandbox',
-    '--disable-speech-api',
-    '--disable-sync',
-    '--hide-scrollbars',
-    '--ignore-gpu-blacklist',
-    '--metrics-recording-only',
-    '--mute-audio',
-    '--no-default-browser-check',
-    '--no-first-run',
-    '--no-pings',
-    '--no-sandbox',
-    '--no-zygote',
-    '--password-store=basic',
-    '--use-gl=swiftshader',
-    '--use-mock-keychain',
-    '--fast-start',
-  ],
-}
