--- lib/bdo_codex/index.mjs
@@ -1,2 +1,0 @@
-export { searchCodexForRecipes } from './searchCodex.mjs'
-export { getItemCodexData } from './getItemCodexData.mjs'
