--- lib/central_market/index.mjs
@@ -1,1 +1,0 @@
-export { getConsumableMarketData } from './getMarketData.mjs'
