--- lib/recipe/recipe_logger.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class RecipeLogger
-end
