--- utils/npc_item_index.rb
@@ -1,515 +1,0 @@
-# frozen_string_literal: true
-
-class NPCItemIndex
-  def self.get_item(id)
-    ITEMS.dig(id.to_i)
-  end
-
-  ITEMS = {
-    7313 => {
-    name: 'Apple',
-    id: 7313,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    7316 => {
-    name: 'Banana',
-    id: 7316,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    7314 => {
-    name: 'Cherry',
-    id: 7314,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    7312 => {
-    name: 'Paprika',
-    id: 7312,
-    price: 850,
-    price_per_one: 850,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    7315 => {
-    name: 'Pear',
-    id: 7315,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    7317 => {
-    name: 'Pineapple',
-    id: 7317,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    7304 => {
-    name: 'Strawberry',
-    id: 9008,
-    price: 700,
-    price_per_one: 700,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'fruit',
-    main_category: 25,
-    sub_category: 3,
-  },
-    9016 => {
-    name: 'Deep Frying Oil',
-    id: 9016,
-    price: 40,
-    price_per_one: 40,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    9015 => {
-    name: 'Olive Oil',
-    id: 9015,
-    price: 40,
-    price_per_one: 40,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    9007 => {
-    name: 'Cooking Oil',
-    id: 9007,
-    price: 20,
-    price_per_one: 20,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    9008 => {
-    name: 'Sun-Dried Salt',
-    id: 9008,
-    price: 200,
-    price_per_one: 200,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    9002 => {
-    name: 'Sugar',
-    id: 9002,
-    price: 20,
-    price_per_one: 20,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    9001 => {
-    name: 'Salt',
-    id: 9001,
-    price: 20,
-    price_per_one: 20,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    9059 => {
-    name: 'Mineral Water',
-    id: 9059,
-    price: 30,
-    price_per_one: 30,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'cooking',
-  },
-    6656 => {
-    name: 'Purified Water',
-    id: 6656,
-    price: 5000,
-    price_per_one: 5000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'desert goods',
-    main_category: 25,
-    sub_category: 8,
-  },
-    8041 => {
-    name: 'Old Moon Solvent',
-    id: 8041,
-    price: 100000,
-    price_per_one: 100000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'old moon',
-  },
-    820936 => {
-    name: 'Spellbound Catalyst',
-    id: 820936,
-    price: 1000000,
-    price_per_one: 1000000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'old moon',
-  },
-    59378 => {
-    name: 'Growth Reagent',
-    id: 59378,
-    price: 60000000,
-    price_per_one: 60000000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'old moon',
-  },
-    4986 => {
-    name: "Ibellab's Essence",
-    id: 4986,
-    price: 10000,
-    price_per_one: 10000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'old moon',
-  },
-    4916 => {
-    name: 'Blue Reagent',
-    id: 4916,
-    price: 50000,
-    price_per_one: 50000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'material vendor',
-  },
-    4925 => {
-    name: 'Tears of the Falling Moon',
-    id: 4925,
-    price: 10000,
-    price_per_one: 10000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'old moon',
-  },
-    45162 => {
-    name: 'Old Moon Alchemy Catalyst',
-    id: 45162,
-    price: 3500000,
-    price_per_one: 3500000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'old moon',
-  },
-    4914 => {
-    name: "Aal's Tear",
-    id: 4914,
-    price: 21000,
-    price_per_one: 21000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'material vendor',
-  },
-    5408 => {
-    name: 'Arrow Mushroom',
-    id: 5408,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5410 => {
-    name: 'Cloud Mushroom',
-    id: 5410,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5409 => {
-    name: 'Dwarf Mushroom',
-    id: 5409,
-    price: 950,
-    price_per_one: 950,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5413 => {
-    name: 'Emperor Mushroom',
-    id: 5413,
-    price: 900,
-    price_per_one: 900,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5407 => {
-    name: 'Fortune Teller Mushroom',
-    id: 5407,
-    price: 900,
-    price_per_one: 900,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5414 => {
-    name: 'Ghost Mushroom',
-    id: 5414,
-    price: 1000,
-    price_per_one: 1000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5411 => {
-    name: 'Sky Mushroom',
-    id: 5411,
-    price: 950,
-    price_per_one: 950,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5412 => {
-    name: 'Tiger Mushroom',
-    id: 5412,
-    price: 950,
-    price_per_one: 950,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'mushroom vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5404 => {
-    name: 'Dry Mane Grass',
-    id: 5404,
-    price: 525,
-    price_per_one: 525,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'herb vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5403 => {
-    name: 'Fire Flake Flower',
-    id: 5403,
-    price: 550,
-    price_per_one: 550,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'herb vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5405 => {
-    name: 'Silk Honey Grass',
-    id: 5405,
-    price: 500,
-    price_per_one: 500,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'herb vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5402 => {
-    name: 'Silver Azalea',
-    id: 5402,
-    price: 575,
-    price_per_one: 575,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'herb vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    5401 => {
-    name: 'Sunrise Herb',
-    id: 5401,
-    price: 500,
-    price_per_one: 500,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'herb vendor',
-    main_category: 25,
-    sub_category: 2,
-  },
-    15009 => {
-    name: 'Magic Crystal of Abundance - Healing',
-    id: 15009,
-    price: 30000,
-    price_per_one: 30000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'jeweler',
-  },
-    15001 => {
-    name: 'Magic Crystal of Flame - Precision',
-    id: 15001,
-    price: 30000,
-    price_per_one: 30000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'jeweler',
-  },
-    15020 => {
-    name: 'Magic Crystal of Enchantment - Agility',
-    id: 15020,
-    price: 30000,
-    price_per_one: 30000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'jeweler',
-  },
-    15019 => {
-    name: 'Magic Crystal of Enchantment - Vision',
-    id: 15019,
-    price: 30000,
-    price_per_one: 30000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'jeweler',
-  },
-    15013 => {
-    name: 'Magic Crystal of Nature - Ascension',
-    id: 15013,
-    price: 30000,
-    price_per_one: 30000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'jeweler',
-  },
-    4915 => {
-    name: 'Manos Stone',
-    id: 4915,
-    price: 5000000,
-    price_per_one: 5000000,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'jeweler',
-  },
-    4903 => {
-    name: 'Empty Crystal',
-    id: 4903,
-    price: 2500,
-    price_per_one: 2500,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-    main_category: 25,
-    sub_category_: 8,
-  },
-    517 => {
-    name: 'HP Potion (Small)',
-    id: 517,
-    price: 80,
-    price_per_one: 80,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-    main_category: 35,
-    sub_category_: 5,
-  },
-    521 => {
-    name: 'HP Potion (Medium)',
-    id: 517,
-    price: 200,
-    price_per_one: 200,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-    main_category: 35,
-    sub_category_: 5,
-  },
-    519 => {
-    name: 'HP Potion (Large)',
-    id: 517,
-    price: 500,
-    price_per_one: 500,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-    main_category: 35,
-    sub_category_: 5,
-  },
-    524 => {
-    name: 'HP Potion (Extra Large)',
-    id: 524,
-    price: 1250,
-    price_per_one: 1250,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-    main_category: 35,
-    sub_category_: 5,
-  },
-    520 => {
-    name: 'MP Potion (Small)',
-    id: 520,
-    price: 80,
-    price_per_one: 80,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-    main_category: 35,
-    sub_category_: 5,
-  },
-    522 => {
-    name: 'MP Potion (Large)',
-    id: 522,
-    price: 500,
-    price_per_one: 500,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-  },
-    526 => {
-    name: 'MP Potion (Extra Large)',
-    id: 526,
-    price: 1250,
-    price_per_one: 1250,
-    count: Float::INFINITY,
-    is_npc_item: true,
-    npc_type: 'general goods',
-  },
-  }
-
-end
