--- utils/recipe_utils/index.mjs
@@ -1,3 +1,0 @@
-export { NPC_ITEM_INDEX } from './npcItemList.mjs'
-export { logRecipeInfo } from './logRecipeInfo.mjs'
-export { getAllRecipePrices } from './getAllRecipePrices.mjs'
