--- utils/recipe_utils/npcItemList.mjs
@@ -1,410 +1,0 @@
-export const NPC_ITEM_INDEX = {
-  7313: {
-    name: 'Apple',
-    mainKey: 7313,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  7316: {
-    name: 'Banana',
-    mainKey: 7316,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  7314: {
-    name: 'Cherry',
-    mainKey: 7314,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  7312: {
-    name: 'Paprika',
-    mainKey: 7312,
-    pricePerOne: 850,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  7315: {
-    name: 'Pear',
-    mainKey: 7315,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  7317: {
-    name: 'Pineapple',
-    mainKey: 7317,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  7304: {
-    name: 'Strawberry',
-    mainKey: 9008,
-    pricePerOne: 700,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'fruit',
-  },
-  9016: {
-    name: 'Deep Frying Oil',
-    mainKey: 9016,
-    pricePerOne: 40,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  9015: {
-    name: 'Olive Oil',
-    mainKey: 9015,
-    pricePerOne: 40,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  9007: {
-    name: 'Cooking Oil',
-    mainKey: 9007,
-    pricePerOne: 20,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  9008: {
-    name: 'Sun-Dried Salt',
-    mainKey: 9008,
-    pricePerOne: 200,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  9002: {
-    name: 'Sugar',
-    mainKey: 9002,
-    pricePerOne: 20,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  9001: {
-    name: 'Salt',
-    mainKey: 9001,
-    pricePerOne: 20,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  9059: {
-    name: 'Mineral Water',
-    mainKey: 9059,
-    pricePerOne: 30,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'cooking',
-  },
-  6656: {
-    name: 'Purified Water',
-    mainKey: 6656,
-    pricePerOne: 5000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'desert goods',
-  },
-  8041: {
-    name: 'Old Moon Solvent',
-    mainKey: 8041,
-    pricePerOne: 100000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'old moon',
-  },
-  820936: {
-    name: 'Spellbound Catalyst',
-    mainKey: 820936,
-    pricePerOne: 1000000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'old moon',
-  },
-  59378: {
-    name: 'Growth Reagent',
-    mainKey: 59378,
-    pricePerOne: 60000000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'old moon',
-  },
-  4986: {
-    name: "Ibellab's Essence",
-    mainKey: 4986,
-    pricePerOne: 10000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'old moon',
-  },
-  4916: {
-    name: 'Blue Reagent',
-    mainKey: 4916,
-    pricePerOne: 50000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'material vendor',
-  },
-  4925: {
-    name: 'Tears of the Falling Moon',
-    mainKey: 4925,
-    pricePerOne: 10000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'old moon',
-  },
-  45162: {
-    name: 'Old Moon Alchemy Catalyst',
-    mainKey: 45162,
-    pricePerOne: 3500000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'old moon',
-  },
-  4914: {
-    name: "Aal's Tear",
-    mainKey: 4914,
-    pricePerOne: 21000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'material vendor',
-  },
-  5408: {
-    name: 'Arrow Mushroom',
-    mainKey: 5408,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5410: {
-    name: 'Cloud Mushroom',
-    mainKey: 5410,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5409: {
-    name: 'Dwarf Mushroom',
-    mainKey: 5409,
-    pricePerOne: 950,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5413: {
-    name: 'Emperor Mushroom',
-    mainKey: 5413,
-    pricePerOne: 900,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5407: {
-    name: 'Fortune Teller Mushroom',
-    mainKey: 5407,
-    pricePerOne: 900,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5414: {
-    name: 'Ghost Mushroom',
-    mainKey: 5414,
-    pricePerOne: 1000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5411: {
-    name: 'Sky Mushroom',
-    mainKey: 5411,
-    pricePerOne: 950,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5412: {
-    name: 'Tiger Mushroom',
-    mainKey: 5412,
-    pricePerOne: 950,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'mushroom vendor',
-  },
-  5404: {
-    name: 'Dry Mane Grass',
-    mainKey: 5404,
-    pricePerOne: 525,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'herb vendor',
-  },
-  5403: {
-    name: 'Fire Flake Flower',
-    mainKey: 5403,
-    pricePerOne: 550,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'herb vendor',
-  },
-  5405: {
-    name: 'Silk Honey Grass',
-    mainKey: 5405,
-    pricePerOne: 500,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'herb vendor',
-  },
-  5402: {
-    name: 'Silver Azalea',
-    mainKey: 5402,
-    pricePerOne: 575,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'herb vendor',
-  },
-  5401: {
-    name: 'Sunrise Herb',
-    mainKey: 5401,
-    pricePerOne: 500,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'herb vendor',
-  },
-  15009: {
-    name: 'Magic Crystal of Abundance - Healing',
-    mainKey: 15009,
-    pricePerOne: 30000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'jeweler',
-  },
-  15001: {
-    name: 'Magic Crystal of Flame - Precision',
-    mainKey: 15001,
-    pricePerOne: 30000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'jeweler',
-  },
-  15020: {
-    name: 'Magic Crystal of Enchantment - Agility',
-    mainKey: 15020,
-    pricePerOne: 30000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'jeweler',
-  },
-  15019: {
-    name: 'Magic Crystal of Enchantment - Vision',
-    mainKey: 15019,
-    pricePerOne: 30000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'jeweler',
-  },
-  15013: {
-    name: 'Magic Crystal of Nature - Ascension',
-    mainKey: 15013,
-    pricePerOne: 30000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'jeweler',
-  },
-  4915: {
-    name: 'Manos Stone',
-    mainKey: 4915,
-    pricePerOne: 5000000,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'jeweler',
-  },
-  4903: {
-    name: 'Empty Crystal',
-    mainKey: 4903,
-    pricePerOne: 2500,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  517: {
-    name: 'HP Potion (Small)',
-    mainKey: 517,
-    pricePerOne: 80,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  521: {
-    name: 'HP Potion (Medium)',
-    mainKey: 517,
-    pricePerOne: 200,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  519: {
-    name: 'HP Potion (Large)',
-    mainKey: 517,
-    pricePerOne: 500,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  519: {
-    name: 'HP Potion (Extra Large)',
-    mainKey: 524,
-    pricePerOne: 1250,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  520: {
-    name: 'MP Potion (Small)',
-    mainKey: 520,
-    pricePerOne: 80,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  521: {
-    name: 'MP Potion (Medium)',
-    mainKey: 517,
-    pricePerOne: 200,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  522: {
-    name: 'MP Potion (Large)',
-    mainKey: 522,
-    pricePerOne: 500,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-  526: {
-    name: 'MP Potion (Extra Large)',
-    mainKey: 526,
-    pricePerOne: 1250,
-    count: Infinity,
-    isNPCItem: true,
-    npcType: 'general goods',
-  },
-}
